+++ src/Components/API/addressAutocomplete.ts
@@ -17,18 +17,3 @@ export type CompleteAddress = AddressSuggestion & {
   postalCode: string
   city: string
 }
-
-const MANDATORY_ADDRESS_FIELDS: Array<keyof AddressSuggestion> = [
-  'id',
-  'address',
-  'streetName',
-  'streetNumber',
-  'postalCode',
-  'city',
-]
-
-export const isCompleteAddress = (
-  suggestion: AddressSuggestion,
-): suggestion is CompleteAddress => {
-  return MANDATORY_ADDRESS_FIELDS.every((key) => suggestion[key] !== undefined)
-}
