+++ src/Components/Actions/AddressAutocompleteAction/AddressAutocompleteAction.tsx
@@ -6,22 +6,19 @@ import { Card as BaseCard, Input, Spacer } from '../Common'
 import styled from '@emotion/styled'
 import { ApiContext } from '../../API/ApiContext'
 import { ApiComponent } from '../../API/apiComponent'
-import { useCombobox } from 'downshift'
 import {
   AddressSuggestion,
   CompleteAddress,
-  isCompleteAddress,
 } from '../../API/addressAutocomplete'
 import { colorsV3, fonts } from '@hedviginsurance/brand'
 import { ContinueButton } from '../../ContinueButton'
-import Modal from './Modal'
-import useAddressSearch from './useAddressSearch'
 import {
-  isMatchingStreetName,
   formatAddressLine,
   formatAddressLines,
+  isCompleteAddress,
+  isSameAddress,
 } from './utils'
-import { Cross } from '../../Icons/Cross'
+import AddressAutocomplete from './AddressAutocomplete'
 
 const ADDRESS_NOT_FOUND = 'ADDRESS_NOT_FOUND'
 
@@ -70,198 +67,19 @@ const FakeInput = styled(Input)`
   }
 `
 
-const ModalHeader = styled.header`
-  box-sizing: border-box;
-  width: 100%;
-  flex-shrink: 0;
-
-  padding: 0 16px 16px;
-  border-bottom: 1px solid ${colorsV3.gray300};
-`
-
-const ModalHeaderRow = styled.div<{ align: 'center' | 'flex-start' }>`
-  height: 56px;
-  display: flex;
-  align-items: ${(props) => props.align};
-  justify-content: flex-end;
-`
-
-const ModalHeaderLabel = styled.label`
-  position: absolute;
-  left: 0;
-  right: 0;
-  z-index: 1;
-
-  color: ${colorsV3.black};
-  font-family: ${fonts.FAVORIT};
-  text-align: center;
-  font-size: 16px;
-`
-
-const ModalHeaderButton = styled.button`
-  border: 0;
-  background: transparent;
-  appearance: none;
-  outline: 0;
-  position: relative;
-  z-index: 2;
-
-  color: ${colorsV3.black};
-  font-family: ${fonts.FAVORIT};
-  font-size: 16px;
-
-  &:focus {
-    border-radius: 4px;
-    box-shadow: 0 0 0 3px ${colorsV3.purple300};
-  }
-`
-
-const ComboboxField = styled.div`
-  box-sizing: border-box;
-  text-align: left;
-  width: 100%;
-  position: relative;
-
-  padding: 8px 16px;
-  border-radius: 8px;
-  border: 1px solid ${colorsV3.gray500};
-
-  &:focus-within {
-    border-color: ${colorsV3.purple500};
-    box-shadow: 0 0 0 2px ${colorsV3.purple300};
-  }
-`
-
-const ClearButton = styled.button`
-  position: absolute;
-  right: 16px;
-  top: calc(50% - 9px);
-  cursor: pointer;
-  height: 18px;
-  width: 18px;
-  border: 0;
-  border-radius: 9px;
-  background-color: ${colorsV3.gray500};
-  outline: 0;
-  padding: 0;
-
-  &:hover {
-    background-color: ${colorsV3.gray700};
-  }
-
-  &:focus {
-    box-shadow: 0 0 0 3px ${colorsV3.purple300};
-  }
-
-  svg {
-    width: 60%;
-    height: 60%;
-  }
-`
-
-const ComboboxList = styled.ul`
-  padding: 0;
-  margin: 0;
-  width: 100%;
-
-  flex: 1;
-  overflow: auto;
-`
-
-const ComboboxInput = styled.input`
-  width: 100%;
-
-  background: none;
-  border: none;
-  box-sizing: border-box;
-  text-align: left;
-  appearance: none;
-  -moz-appearance: textfield;
-  outline: 0;
-
-  font-family: ${fonts.FAVORIT};
-  font-size: 20px;
-  color: ${colorsV3.black};
-
-  &::placeholder {
-    color: ${colorsV3.gray300};
-  }
-`
-
-const ComboboxOption = styled.li`
-  padding: 8px 16px;
-  min-height: 32px;
-  display: flex;
-  align-items: center;
-
-  text-align: left;
-  font-family: ${fonts.FAVORIT}, sans-serif;
-  font-size: 16px;
-  color: ${colorsV3.gray900};
-  cursor: pointer;
-
-  &:not(:first-of-type) {
-    border-top: 1px solid ${colorsV3.gray300};
-  }
-
-  &[data-highlighted] {
-    background-color: ${colorsV3.purple300};
-    border-top-color: ${colorsV3.purple300};
-  }
-
-  &[data-highlighted] + & {
-    border-top-color: ${colorsV3.purple300};
-  }
-
-  &:active {
-    background-color: ${colorsV3.purple500};
-    border-top-color: ${colorsV3.purple500};
-  }
-
-  &:active + & {
-    border-top-color: ${colorsV3.purple500};
-  }
-`
-
-const ComboboxOptionNotFound = styled(ComboboxOption)`
-  color: ${colorsV3.red500};
-
-  &[data-highlighted] {
-    background-color: ${colorsV3.red500};
-    border-top-color: ${colorsV3.red500};
-    color: ${colorsV3.white};
-  }
-`
-
 const PostalAddress = styled.p`
   font-family: ${fonts.FAVORIT}, sans-serif;
   font-size: 16px;
   text-align: left;
-  margin: 0;
-
   color: ${colorsV3.gray700};
+  margin: 0;
+  padding-left: 16px;
 
-  ${FakeInput} + & {
-    padding-left: 16px;
-
-    @media (min-width: 600px) {
-      font-size: 24px;
-      padding-left: 32px;
-    }
+  @media (min-width: 600px) {
+    font-size: 24px;
+    padding-left: 32px;
   }
 `
-
-const AddressOption: React.FC<{ address: AddressSuggestion }> = ({
-  address,
-}) => {
-  const [addressLine, postalLine] = formatAddressLines(address)
-  return (
-    <div>
-      {addressLine}
-      {postalLine ? <PostalAddress>{postalLine}</PostalAddress> : null}
-    </div>
-  )
-}
 
 export interface AddressAutocompleteActionProps {
   isTransitioning: boolean
@@ -319,19 +137,12 @@ export const AddressAutocompleteAction: React.FC<AddressAutocompleteActionProps>
   props,
 ) => {
   const api = React.useContext(ApiContext)
-  const inputRef = React.useRef<HTMLInputElement>(null)
 
   const [isHovered, setIsHovered] = React.useState(false)
   const { store, setValue, removeValues } = React.useContext(StoreContext)
 
-  const [isModalOpen, setIsModalOpen] = React.useState(false)
-  const handleDismissModal = React.useCallback(() => setIsModalOpen(false), [])
-  React.useEffect(() => {
-    if (isModalOpen) {
-      // Move focus to input on next render
-      setTimeout(() => inputRef.current?.focus(), 1)
-    }
-  }, [isModalOpen])
+  const [isAutocompleteActive, setIsAutocompleteActive] = React.useState(false)
+  const [searchTerm, setSearchTerm] = React.useState('')
 
   const [
     pickedSuggestion,
@@ -343,18 +154,38 @@ export const AddressAutocompleteAction: React.FC<AddressAutocompleteActionProps>
     setConfirmedAddress,
   ] = React.useState<CompleteAddress | null>(null)
 
-  const [textValue, setTextValue] = React.useState('')
+  const confirmSuggestion = React.useCallback(
+    async (
+      newSuggestion: AddressSuggestion,
+      prevSuggestion: AddressSuggestion | null,
+    ) => {
+      if (!isCompleteAddress(newSuggestion)) return null
 
-  const [suggestions, setSuggestions] = useAddressSearch(
-    textValue,
-    pickedSuggestion ?? undefined,
+      if (newSuggestion.floor && newSuggestion.apartment) return newSuggestion
+
+      if (prevSuggestion && isSameAddress(prevSuggestion, newSuggestion))
+        return newSuggestion
+
+      const results = await api.addressAutocompleteQuery(newSuggestion.address)
+      if (results.length === 1) return newSuggestion
+
+      return null
+    },
+    [api],
   )
 
-  const handleClearInput = React.useCallback(() => {
-    setTextValue('')
-    setPickedSuggestion(null)
-    inputRef.current?.focus()
-  }, [])
+  const handleSelectSuggestion = React.useCallback(
+    async (suggestion: AddressSuggestion | null) => {
+      setPickedSuggestion(suggestion)
+
+      if (suggestion) {
+        setConfirmedAddress(
+          await confirmSuggestion(suggestion, pickedSuggestion),
+        )
+      }
+    },
+    [confirmSuggestion, pickedSuggestion],
+  )
 
   const clearStoreValues = React.useCallback(
     () =>
@@ -389,111 +220,30 @@ export const AddressAutocompleteAction: React.FC<AddressAutocompleteActionProps>
     ],
   )
 
+  const handleNoAddressFound = React.useCallback(() => {
+    setIsAutocompleteActive(false)
+    clearStoreValues()
+
+    setValue(STORE_KEY.ADDRESS_SEARCH_TERM, searchTerm)
+    setValue(props.storeKey, ADDRESS_NOT_FOUND)
+    props.onContinue()
+  }, [clearStoreValues, setValue, props.storeKey, searchTerm])
+
   React.useEffect(() => {
     if (confirmedAddress) {
-      inputRef.current?.blur()
-      setIsModalOpen(false)
-      setSuggestions(null)
+      setIsAutocompleteActive(false)
     }
   }, [confirmedAddress])
 
-  const handleNoAddressFound = React.useCallback(() => {
-    setIsModalOpen(false)
-    clearStoreValues()
-
-    setValue(STORE_KEY.ADDRESS_SEARCH_TERM, textValue)
-    setValue(props.storeKey, ADDRESS_NOT_FOUND)
-    props.onContinue()
-  }, [clearStoreValues, setValue, props.storeKey, textValue])
-
-  const comboboxItems = React.useMemo<AddressSuggestion[]>(
-    () => (suggestions ? [...suggestions, { address: ADDRESS_NOT_FOUND }] : []),
-    [suggestions],
-  )
-
-  const {
-    getMenuProps,
-    getInputProps,
-    getComboboxProps,
-    highlightedIndex,
-    getItemProps,
-  } = useCombobox<AddressSuggestion>({
-    selectedItem: pickedSuggestion,
-    inputValue: textValue,
-    items: comboboxItems,
-    itemToString: (item) => (item ? formatAddressLine(item) : ''),
-    onInputValueChange: ({ inputValue, type }) => {
-      if (type === '__input_keydown_escape__') return
-
-      setTextValue(inputValue || '')
-      setConfirmedAddress(null)
-
-      if (!inputValue) {
-        // Reset picked suggestion for empty input field
-        setPickedSuggestion(null)
-      }
-    },
-    onSelectedItemChange: ({ selectedItem }) => {
-      if (selectedItem?.address === ADDRESS_NOT_FOUND) {
-        handleNoAddressFound()
-      }
-
-      // Reset list of suggestions
-      setSuggestions(null)
-      inputRef.current?.focus()
-      setPickedSuggestion(selectedItem ?? null)
-    },
-  })
-
-  React.useEffect(() => {
-    const checkPickedSuggestion = async (suggestion: AddressSuggestion) => {
-      const newSuggestions = await api.addressAutocompleteQuery(
-        suggestion.address,
-      )
-      const oneResultLeft = newSuggestions.length === 1
-      const sameResultsAsBefore = newSuggestions.every(
-        (newSuggestion, index) => newSuggestion.id === suggestions?.[index]?.id,
-      )
-      if (
-        (oneResultLeft || sameResultsAsBefore) &&
-        isCompleteAddress(suggestion)
-      ) {
-        setConfirmedAddress(suggestion)
-      }
-    }
-
-    if (pickedSuggestion && pickedSuggestion.id) {
-      if (
-        isCompleteAddress(pickedSuggestion) &&
-        pickedSuggestion.floor &&
-        pickedSuggestion.apartment
-      ) {
-        setConfirmedAddress(pickedSuggestion)
-      } else {
-        checkPickedSuggestion(pickedSuggestion)
-      }
-    } else {
-      setConfirmedAddress(null)
-    }
-
-    return () => setConfirmedAddress(null)
-  }, [pickedSuggestion])
-
-  const [pickedAddressLine, pickedPostalLine] = React.useMemo(() => {
-    if (pickedSuggestion) {
-      const [addressLine, postalLine] = formatAddressLines(pickedSuggestion)
-      const isMatching = isMatchingStreetName(textValue, pickedSuggestion)
-      return [addressLine, isMatching ? postalLine : undefined]
-    }
-
-    return []
-  }, [pickedSuggestion, textValue])
+  const [confirmedAddressLine, confirmedPostalLine] = confirmedAddress
+    ? formatAddressLines(confirmedAddress)
+    : []
 
   return (
     <Container>
       <motion.div
         animate={{
-          opacity: isModalOpen ? 0 : 1,
+          opacity: isAutocompleteActive ? 0 : 1,
         }}
         transition={{ ease: 'easeOut', duration: 0.25 }}
       >
@@ -507,13 +257,13 @@ export const AddressAutocompleteAction: React.FC<AddressAutocompleteActionProps>
 
           <FakeInput
             placeholder={props.placeholder}
-            value={pickedAddressLine || textValue}
-            onClick={() => setIsModalOpen(true)}
-            onFocus={() => setIsModalOpen(true)}
-            size={(pickedAddressLine || textValue).length}
+            value={confirmedAddressLine}
+            onClick={() => setIsAutocompleteActive(true)}
+            onFocus={() => setIsAutocompleteActive(true)}
+            size={(confirmedAddressLine || props.placeholder).length}
           />
-          {pickedPostalLine ? (
-            <PostalAddress>{pickedPostalLine}</PostalAddress>
+          {confirmedPostalLine ? (
+            <PostalAddress>{confirmedPostalLine}</PostalAddress>
           ) : null}
         </Card>
       </motion.div>
@@ -523,63 +273,19 @@ export const AddressAutocompleteAction: React.FC<AddressAutocompleteActionProps>
       <ContinueButton
         onClick={() => handleContinue(confirmedAddress!)}
         disabled={!confirmedAddress}
-        text={(props.link || {}).label || 'Nästa'}
+        text={(props.link || {}).label || 'Continue'}
       />
 
-      <Modal isOpen={isModalOpen} onDismiss={handleDismissModal}>
-        <ModalHeader>
-          <ModalHeaderRow align="center">
-            <ModalHeaderLabel>Address</ModalHeaderLabel>
-            <ModalHeaderButton onClick={handleDismissModal}>
-              Cancel
-            </ModalHeaderButton>
-          </ModalHeaderRow>
-
-          <ComboboxField {...getComboboxProps()}>
-            <ComboboxInput
-              {...getInputProps({
-                ref: inputRef,
-                placeholder: props.placeholder,
-              })}
-            />
-            {pickedPostalLine ? (
-              <PostalAddress>{pickedPostalLine}</PostalAddress>
-            ) : null}
-
-            {textValue ? (
-              <ClearButton onClick={handleClearInput}>
-                <Cross />
-              </ClearButton>
-            ) : null}
-          </ComboboxField>
-        </ModalHeader>
-
-        <ComboboxList {...getMenuProps()}>
-          {comboboxItems.map((item, index) => {
-            const itemProps = {
-              key: `${item.address}${index}`,
-              ...(highlightedIndex === index && {
-                'data-highlighted': true,
-              }),
-              ...getItemProps({ item, index }),
-            }
-
-            if (item.address === ADDRESS_NOT_FOUND) {
-              return (
-                <ComboboxOptionNotFound {...itemProps}>
-                  Can't find my address
-                </ComboboxOptionNotFound>
-              )
-            }
-
-            return (
-              <ComboboxOption {...itemProps}>
-                <AddressOption address={item} />
-              </ComboboxOption>
-            )
-          })}
-        </ComboboxList>
-      </Modal>
+      <AddressAutocomplete
+        isActive={isAutocompleteActive}
+        onDismiss={() => setIsAutocompleteActive(false)}
+        selected={pickedSuggestion}
+        onSelect={handleSelectSuggestion}
+        onNotFound={handleNoAddressFound}
+        value={searchTerm}
+        onChange={setSearchTerm}
+        placeholder={props.placeholder}
+      />
     </Container>
   )
 }
